junos: refactor RPC requests/responses
- split constants and structs into dedicated file
- rename some constants and structs
- fix xml.Name on structs
